Extract status action
